+++ prisma/migrations/20260916164740_add_beta_signup_invited_at/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "BetaSignup" ADD COLUMN "invitedAt" DATETIME;